decimal to double amount_of_itmes

+++ database/migrations/2025_10_12_205257_modify_amount_of_items_in_order_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('order_items', function (Blueprint $table) {
-            $table->decimal('amount_of_items', 10)->change();
+            $table->double('amount_of_items')->change();
         });
     }
